+++ CHANGELOG.rst
@@ -43,3 +43,5 @@ Bug Fixes
 - Raise error if a complex spin ``S`` is provided to ``genmagstr`` in
   ``helical`` mode. Previously this meant it would silently ignore the
   ``n`` option, and behave exactly like ``fourier`` mode.
+- Raise error if ``rotate`` mode is used without first initialising
+  a magnetic structure